Integrate new cache with impressions observer

+++ src/main/java/io/split/android/client/service/ServiceConstants.java
@@ -32,7 +32,7 @@ public class ServiceConstants {
     public static final String WORKER_PARAM_CONFIGURED_FILTER_VALUES = "configuredFilterValues";
     public static final String WORKER_PARAM_CONFIGURED_FILTER_TYPE = "configuredFilterType";
 
-    public static final long LAST_SEEN_IMPRESSION_CACHE_SIZE = 500;
+    public static final int LAST_SEEN_IMPRESSION_CACHE_SIZE = 500;
     public static final int MY_SEGMENT_V2_DATA_SIZE = 1024 * 10;// bytes
     public static final long DEFAULT_SSE_CONNECTION_DELAY_SECS = 60;
